[a16wmix cleanup] Remove a4w4/a8w4 artifacts; consolidate common+host into gemm1+__init__
Task 1: delete docs/issues/mxfp_moe_a4w4_a8w4_broken.md and
tools/repro/repro_mxfp_moe_a4w4_a8w4_broken.py; replace their references in the
test_moe_gemm.py a4w4/a8w4 xfail messages with a brief inline note (xfails kept).

Task 2: consolidate the moe_2stage_a16wmix package from 5 files to 3.
  - common.py -> gemm1.py (shared low-level helpers; gemm2 imports them from .gemm1,
    which it already did for A16WI4_GROUP_SIZE et al -- no circular import).
  - host.py  -> __init__.py (launch glue + config resolvers; __init__ already
    re-exported these as the public API).
Kernel bodies byte-identical (pure code move); all load-bearing comments preserved.
