+++ kernels/moe/moe_2stage_a16wmix/gemm1.py
@@ -1,6 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright (C) 2025-2026 FlyDSL Project Contributors
 
+import os
+
 import flydsl.compiler as flyc
 import flydsl.expr as fx
 from flydsl._mlir import ir
@@ -8,24 +10,195 @@ from flydsl._mlir.dialects import llvm
 from flydsl.expr import arith, const_expr, gpu, range_constexpr, rocdl
 from flydsl.expr.typing import T
 from flydsl.expr.typing import Vector as Vec
+from flydsl.runtime.device import get_rocm_arch
 from kernels.common import buffer_ops
 from kernels.common.layout_utils import crd2idx
 
-from .common import (
-    _buffer_rsrc,
-    _gep3,
-    _global_i32_at,
-    _global_i32_buffer_tiles,
-    _global_i32_buffer_view,
-    _int4_nibble_to_bf16x8,
-    _lds_ptr3,
-    _raw,
-    _silu_mul_batch,
-    _situ_mul_batch,
-    _udiv,
-    _umod,
-    a16wmix_use_k16,
-)
+# =============================================================================
+# Shared low-level helpers for the a16w4/a16wi4/a16w16 fused MoE kernels. Used by
+# both stage1 (this module) and stage2 (:mod:`gemm2`, imported from here).
+# =============================================================================
+
+_PTR3 = "!llvm.ptr<3>"
+LOG2E = 1.4426950408889634
+
+
+def a16wmix_use_k16(arch=None):
+    """True for the gfx942 (CDNA3) codepath: K=16 MFMA + scalar int4 dequant.
+
+    Arch-gate: gfx950 (CDNA4) has K=32 mfma_f32_16x16x32_bf16 + v_cvt_pk_bf16_f32;
+    gfx942 has neither and falls back to K=16 MFMA + scalar-trunc dequant.
+    ``FLYDSL_A16WMIX_FORCE_K16=1`` forces the gfx942 path (a strict ISA subset) for
+    validation on a gfx950 box.
+    """
+    if os.environ.get("FLYDSL_A16WMIX_FORCE_K16", "0") not in ("0", "", "false", "False"):
+        return True
+    if arch is None:
+        arch = get_rocm_arch() or ""
+    return "gfx95" not in str(arch)
+
+
+def _raw(v):
+    if not isinstance(v, ir.Value) and hasattr(v, "ir_value"):
+        return v.ir_value()
+    return v
+
+
+def _udiv(a, c):
+    cc = fx.Int32(c) if isinstance(c, int) else c
+    return fx.Int32(arith.divui(_raw(a), _raw(cc)))
+
+
+def _umod(a, c):
+    cc = fx.Int32(c) if isinstance(c, int) else c
+    return fx.Int32(arith.remui(_raw(a), _raw(cc)))
+
+
+def _global_i32_buffer_view(addr_i64, num_bytes):
+    # fx.copy BufferCopy atoms take soffset as an element count (not bytes); the
+    # make_layout dynamic-shape leaf must be i32/i64, not fx.Index.
+    num_bytes_i64 = fx.Int64(num_bytes)
+    ptr_ty = fx.PointerType.get(T.i32, address_space=fx.AddressSpace.Global, alignment=4)
+    ptr = fx.inttoptr(ptr_ty, fx.Int64(addr_i64))
+    view = fx.Tensor(fx.make_view(ptr, fx.make_layout(num_bytes_i64 // fx.Int64(4), 1)))
+    return fx.rocdl.make_buffer_tensor(view, max_size=False, num_records_bytes=num_bytes_i64)
+
+
+def _global_i32_buffer_tiles(addr_i64, num_bytes, tile_elems):
+    return fx.logical_divide(_global_i32_buffer_view(addr_i64, num_bytes), fx.make_layout(tile_elems, 1))
+
+
+def _lds_ptr3(base_i32, byte_off_i32):
+    addr_i64 = fx.Int64(base_i32 + byte_off_i32)
+    return llvm.inttoptr(ir.Type.parse(_PTR3), _raw(addr_i64))
+
+
+def _gep3(base_ptr, byte_off_i32):
+    return buffer_ops.get_element_ptr(base_ptr, byte_offset=_raw(byte_off_i32), elem_type=T.i8)
+
+
+def _global_base_ptr1(addr_i64):
+    return llvm.inttoptr(ir.Type.parse("!llvm.ptr<1>"), _raw(fx.Int64(addr_i64)))
+
+
+def _gep1(base_ptr, byte_off_i32):
+    return buffer_ops.get_element_ptr(base_ptr, byte_offset=_raw(byte_off_i32), elem_type=T.i8)
+
+
+def _global_i32_ptr(addr_i64):
+    ptr_ty = fx.PointerType.get(T.i32, address_space=fx.AddressSpace.Global, alignment=4)
+    return fx.inttoptr(ptr_ty, fx.Int64(addr_i64))
+
+
+def _global_i32_at(addr_i64, idx):
+    return _global_i32_ptr(addr_i64)[idx]
+
+
+def _buffer_rsrc(addr_i64, num_records_bytes):
+    return buffer_ops.create_buffer_resource_from_addr(_raw(fx.Int64(addr_i64)), num_records_bytes=num_records_bytes)
+
+
+def _silu_mul_batch(gs, us):
+    e = [fx.Float32(rocdl.exp2(T.f32, _raw(g * fx.Float32(-LOG2E)))) for g in gs]
+    sig = [fx.Float32(rocdl.rcp(T.f32, _raw(fx.Float32(1.0) + ei))) for ei in e]
+    return [gs[i] * sig[i] * us[i] for i in range(len(gs))]
+
+
+def _sigmoid_f32(g):
+    e = fx.Float32(rocdl.exp2(T.f32, _raw(g * fx.Float32(-LOG2E))))
+    return fx.Float32(rocdl.rcp(T.f32, _raw(fx.Float32(1.0) + e)))
+
+
+def _tanh_f32(x):
+    # tanh via exp2/rcp, sign-restored (aiter mixed_moe tanh_elem):
+    #   t = (1-exp(-2|x|))/(1+exp(-2|x|)),  tanh(x) = sign(x)*t
+    neg_two_log2e = fx.Float32(-2.0 * LOG2E)
+    abs_x = x.maximumf(-x)
+    e = fx.Float32(rocdl.exp2(T.f32, _raw(abs_x * neg_two_log2e)))
+    recip = fx.Float32(rocdl.rcp(T.f32, _raw(fx.Float32(1.0) + e)))
+    tanh_abs = (fx.Float32(1.0) - e) * recip
+    is_pos = arith.cmpf(arith.CmpFPredicate.OGT, _raw(x), _raw(fx.Float32(0.0)))
+    return fx.Float32(arith.select(is_pos, _raw(tanh_abs), _raw(-tanh_abs)))
+
+
+def _situ_mul_batch(gs, us, situ_beta=1.0, situ_linear_beta=1.0, clamp_limit=7.0):
+    """SiTUv2 activation (aiter mixed_moe situ_mul_vec4):
+        situ(g)    = beta * tanh(g / beta) * sigmoid(g)
+        situ_up(u) = linear_beta * tanh(u / linear_beta)
+        out        = situ(clamp_gate(g)) * situ_up(clamp_lin(u))
+    clamp_gate: g <= +limit (upper only); clamp_lin: u in [-limit, +limit].
+    """
+    neg_lim = fx.Float32(-clamp_limit)
+    beta = fx.Float32(situ_beta)
+    beta_rcp = fx.Float32(1.0 / situ_beta)
+    lbeta = fx.Float32(situ_linear_beta)
+    lbeta_rcp = fx.Float32(1.0 / situ_linear_beta)
+
+    out = []
+    for i in range(len(gs)):
+        # clamp_gate: g <= +lim (upper only, via -max(-g,-lim)); clamp_lin: u in [-lim,+lim].
+        g = -((-gs[i]).maximumf(neg_lim))
+        u = (-((-us[i]).maximumf(neg_lim))).maximumf(neg_lim)
+        situ_g = beta * _tanh_f32(g * beta_rcp) * _sigmoid_f32(g)
+        situ_u = lbeta * _tanh_f32(u * lbeta_rcp)
+        out.append(situ_g * situ_u)
+    return out
+
+
+def _cvt_pk_bf16_f32_se(src_a_f32, src_b_f32):
+    # Side-effecting v_cvt_pk_bf16_f32 (pack 2 f32 -> 2xbf16 in i32). LOAD-BEARING:
+    # the stateless rocdl.cvt_pk_bf16_f32 gets CSE-merged/reordered across K steps in
+    # the a16wi4 gemm1 hot loop (garbage output); side_effects pins each call.
+    return llvm.inline_asm(
+        ir.IntegerType.get_signless(32),
+        [_raw(src_a_f32), _raw(src_b_f32)],
+        "v_cvt_pk_bf16_f32 $0, $1, $2",
+        "=v,v,v",
+        has_side_effects=True,
+    )
+
+
+def _int4_nibble_to_bf16x8(raw_i32, scale_f32, *, use_k16=False):
+    """int4 (signed) -> bf16 upconvert for one MFMA K32 step (8 nibbles -> v8bf16).
+
+    ``raw_i32`` holds 8 signed-int4 nibbles in bits[4n+3:4n] (same K order as the
+    mxfp4 sel 0..3 path). ``v_cvt_off_f32_i4`` reads the nibble unsigned, subtracts 8,
+    and scales the mantissa by 16, so the x16 is folded into eff = scale*16.
+    ``use_k16`` (gfx942): v_cvt_pk_bf16_f32 is gfx950-only -> scalar .to(BFloat16).
+    """
+    eff = fx.Float32(scale_f32 * fx.Float32(16.0))
+    raw_even = fx.Int32(raw_i32)
+    raw_odd = raw_even.shrui(fx.Int32(4))
+    if use_k16:
+        # gfx942 fallback: scalar f32 -> bf16 truncation (no v_cvt_pk_bf16_f32).
+        bf16s = []
+        for j in range_constexpr(4):
+            f_lo = fx.Float32(rocdl.cvt_off_f32_i4(_raw(raw_even), byte_sel=j)) * eff
+            f_hi = fx.Float32(rocdl.cvt_off_f32_i4(_raw(raw_odd), byte_sel=j)) * eff
+            bf16s.append(f_lo.to(fx.BFloat16))
+            bf16s.append(f_hi.to(fx.BFloat16))
+        return fx.Vector.from_elements([_raw(x) for x in bf16s], fx.BFloat16)  # v8bf16
+    # byte_sel loads (1 shift total); side-effecting pk-convert.
+    i32s = []
+    for j in range_constexpr(4):
+        f_lo = fx.Float32(rocdl.cvt_off_f32_i4(_raw(raw_even), byte_sel=j)) * eff
+        f_hi = fx.Float32(rocdl.cvt_off_f32_i4(_raw(raw_odd), byte_sel=j)) * eff
+        i32s.append(fx.Int32(_cvt_pk_bf16_f32_se(_raw(f_lo), _raw(f_hi))))
+    v4i32 = fx.Vector.from_elements([_raw(x) for x in i32s], fx.Int32)
+    return v4i32.bitcast(fx.BFloat16)  # v8bf16
+
+
+def kmchunks_for(BM):
+    return BM // 16
+
+
+def lds_acc_bytes_for(rows, BN):
+    return rows * BN * 4
+
+
+# =============================================================================
+# Stage1 (gate+up GEMM + SiLU/SiTUv2)
+# =============================================================================
 
 # a16wi4 (int4 W) groupwise scale: group_size = 32 == one MFMA K32 step (one ku per
 # K-group). Scale packed bf16 pairs (E, N, G//2, 2); even/odd ku selects lo/hi half.
