+++ tests/kernels/test_moe_gemm.py
@@ -2092,9 +2092,7 @@ def _run_mxfp_moe_e2e(
 _MXFP4_FUSED_XFAIL = pytest.mark.xfail(
     reason="pre-existing mxfp4 fused gemm2 + fp8-gemm1 A-path bug, e2e cos ~0.1; "
     "strict gate exposes it, not an a16w4 regression (kernel is also memory-unsafe "
-    "-> run=False to avoid crashing the session). "
-    "See docs/issues/mxfp_moe_a4w4_a8w4_broken.md; repro "
-    "tools/repro/repro_mxfp_moe_a4w4_a8w4_broken.py. TODO(issue #NNN)",
+    "-> run=False to avoid crashing the session). TODO(issue #NNN)",
     strict=False,
     run=False,
 )
@@ -2440,8 +2438,7 @@ def test_moe_gemm_2stage(
                 "pre-existing mxfp4 fused gemm2 + fp8-gemm1 A-path bug, e2e cos ~0.1; "
                 "strict gate exposes it, not an a16w4 regression (kernel also "
                 "memory-unsafe -> not executed here to avoid crashing the session). "
-                "See docs/issues/mxfp_moe_a4w4_a8w4_broken.md; repro "
-                "tools/repro/repro_mxfp_moe_a4w4_a8w4_broken.py. TODO(issue #NNN)"
+                "TODO(issue #NNN)"
             )
         _run_mxfp_moe_e2e(
             tokens=tokens,
